+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] / 2023-10-28
+### Features
+- Application works offline.
+- Application force Revit Login.
+### Application
+- Update `ricaun.Auth.Aps.UI` to 1.0.4
+- Update scope to `OpenId`.
+- Update only with `IsLoggedIn` and same `UserId`.
+
 ## [1.1.3] / 2023-10-16
 ### Features
 - Application show error in log if fail.
@@ -321,6 +330,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [x] TestsFail
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.2.0]: ../../compare/1.1.3...1.2.0
 [1.1.3]: ../../compare/1.1.2...1.1.3
 [1.1.2]: ../../compare/1.1.1...1.1.2
 [1.1.1]: ../../compare/1.1.0...1.1.1